+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 From 2019 onwards, all notable changes to tarpaulin will be documented in this
 file.
 
+## [0.35.5] 2026-06-21
+### Added
+- Cargo 1.94 onward the `CARGO_BIN_EXE_` env vars are now set at test runtime, replicate this when running under tarpaulin #1833
+
+### Fixed
+- Fix report generation when coverage is below threshold #1852
+
 ## [0.35.4] 2026-04-25
 ### Changed
 - Revert gimli to a non-yanked version